refactor(diagnostics): move /healthz to api/tests/ (test-env-gated)
Previously added the healthz probe to the always-on app urlconf
(`/api/healthz/`). That works but isn't the project's pattern for
test-only routes — `/api/tests/*` already exists for login helpers,
state resets, and other test-suite-only endpoints, and the entire
`tests/` urlconf is gated by `isTestEnvironment()` in
`backend/backend/urls.py` so it's physically not registered in prod.

Moving healthz under `api/tests/` aligns with that pattern:
  - `backend/tests/test_health.py` owns the view
  - `backend/tests/urls.py` registers `healthz/`
  - `backend/app/views_main.py` no longer carries the view
  - `backend/backend/urls.py` no longer routes the prod path
  - reporter URL bumped to `/api/tests/healthz/`

Verified: GET /api/tests/healthz/ → 200 (3.91ms db); GET /api/healthz/
→ 404 (gone from prod path).

Also updated the testing skill (`.claude/skills/testing/references
/backend-test-endpoints.md`) with:
  - Diagnostic Endpoints table entry for healthz
  - "Where do test-only routes live?" section explaining the
    `isTestEnvironment()` gate at the urlconf level (the URL pattern
    is the leak, not just the response — never put test-only routes
    in always-on urlconfs).
